# api/openapi.yaml
openapi: 3.0.2
info:
  description: "The Resource Server is IUDX's data store which allows publication,\
    \ subscription and discovery of data.\nFor search and discovery, it allows users\
    \ to search through temporal, geo-based and attribute queries. \nFor publication\
    \ and subscription, it allows users to use AMQP streaming protocol over TLS.\n\
    It enables *Providers* of datasources to publish data as per the IUDX data descriptor.\n\
    It enables *Consumers* of datasources to search and query for data using HTTPs\
    \ APIs.\nIt enables *Subscribers* a.k.a [Streaming Consumer] of datasources to\
    \ stream data using AMQP streaming protocol over TLS.\n"
  title: IUDX Resource Server APIs
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /iudx/v1/adapter/register:
    post:
      deprecated: false
      description: "Register Adapter API allows a provider to register an adapter\
        \ with IUDX Resource Server. Based on the input provided, the API validates\
        \ the access control of the user with IUDX Auth server to allow or deny access\
        \ to the API. \n\n#### Register Adapter\nTo register an adapter, a user must\
        \ request for a consent from the provider, obtain a token from IUDX Auth server\
        \ and then access the API with the following values:\n  - `resourceGroup`:\
        \ Name of the resource group as registered in IUDX catalogue.\n  - `resourceServer`:\
        \ Domain name of the resource server in which the adapter needs to register.\
        \ \n"
      operationId: registeradapter
      parameters:
      - description: IUDX auth token for accessing API and the resource.
        explode: false
        in: header
        name: token
        required: true
        schema:
          maxLength: 512
          pattern: ^[a-zA-Z0-9\/\@\.]*$
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registeradapterrequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/registeradapter'
          description: Successful query
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/adapterBadRequestErrorResponse'
          description: Bad query
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/standardInvalidTokenErrorResponse'
          description: Unauthorized
      summary: Register adapter
      tags:
      - Data Adaptor
  /iudx/v1/adapter/{id}:
    delete:
      deprecated: false
      description: "Delete Adapter API allows a provider to delete a publisher. Once\
        \ this is done all the `Subscribers` a.k.a [Streaming Consumers] and `Consumers`\
        \ of the data with stop reveiving further updates. Based on the input provided,\
        \ the API validates the access control of the user with IUDX Auth server to\
        \ allow or deny access to the API. \n\n#### Delete Adapter\nTo delete an adapter,\
        \ the provider must present the `id` and `token`. Delete API validates the\
        \ request with the Auth server to allow or deny access to the provider.\n\
        \        \n"
      operationId: deleteadaptor
      parameters:
      - description: IUDX auth token for accessing API and the resource.
        explode: false
        in: header
        name: token
        required: true
        schema:
          maxLength: 512
          pattern: ^[a-zA-Z0-9\/\@\.]*$
          type: string
        style: simple
      - description: id of the adapter obtained from Register adapter.
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 512
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteadapter'
          description: Successful query
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/adapterBadRequestErrorResponse'
          description: Bad query
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/standardInvalidTokenErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/standardResourceNotFoundErrorResponse'
          description: Resource not found
      summary: Delete adaptor
      tags:
      - Data Adaptor
      x-codeSamples:
      - lang: by id
        source: |
          https://rs.iudx.io/iudx/v1/adapter/iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms --header 'token: auth.iudx.org.in/vasanth.rajaraman@datakaveri.org/896f2d9f8b866671a4341b882e3ba2b3'
    get:
      deprecated: false
      description: "Retrieve Adapter API allows a provider to fetch the list of `Subscribers`\
        \ a.k.a [Streaming Consumers] for the data. Based on the input provided, the\
        \ API validates the access control of the user with IUDX Auth server to allow\
        \ or deny access to the API. \n\n#### Retrieve Adapter Details\nTo retrieve\
        \ subscriber list, the provider must present the `id` and `token`. Retrieve\
        \ API validates the request with the Auth server to allow or deny access to\
        \ the provider.\n"
      operationId: getadaptordetails
      parameters:
      - description: IUDX auth token for accessing API and the resource.
        explode: false
        in: header
        name: token
        required: true
        schema:
          maxLength: 512
          pattern: ^[a-zA-Z0-9\/\@\.]*$
          type: string
        style: simple
      - description: id of the adapter obtained from Register adapter.
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 512
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/retrieveadapter'
          description: Successful query
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/adapterBadRequestErrorResponse'
          description: Bad query
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/standardInvalidTokenErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/standardResourceNotFoundErrorResponse'
          description: Resource not found
      summary: Retrieve adapter details
      tags:
      - Data Adaptor
      x-codeSamples:
      - lang: by id
        source: |
          https://rs.iudx.io/iudx/v1/adapter/iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms --header 'token: auth.iudx.org.in/vasanth.rajaraman@datakaveri.org/896f2d9f8b866671a4341b882e3ba2b3'
...
components:
  schemas:
...
    registeradapterrequest:
      example:
        resourceGroup: aqm
        resourceServer: example.com
      properties:
        resourceGroup:
          maxLength: 50
          type: string
        resourceServer:
          maxLength: 50
          type: string
      required:
      - resourceGroup
      - resourceServer
      title: registeradapterrequest
      type: object
...
    registeradapter:
      example:
        username: iudx.io/3ae88e7ba006938ffc38203d326b1711b892226d
        apiKey: 123456
        id: iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms
        URL: databroker.iudx.io
        port: 24567
        vHost: IUDX
      properties:
        username:
          type: string
        apiKey:
          type: string
        id:
          type: string
        URL:
          type: string
        port:
          type: string
        vHost:
          type: string
      required:
      - URL
      - apiKey
      - id
      - port
      - username
      - vHost
      title: registeradapter
      type: object
    retrieveadapter:
      example:
        database:
        - iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms/.*
        adaptorLogs:
        - iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms/.dataIssue
        - iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms/.downstreamIssue
        - iudx.io/aa9d66a000d94a78895de8d4c0b3a67f3450e531/rs.iudx.io/surat-itms/.heartbeat
      properties:
        database:
          items:
            type: string
          type: array
        adaptorLogs:
          items:
            type: string
          type: array
      required:
      - adaptorLogs
      - database
      title: retrieveadapter
      type: object
    deleteadapter:
      example:
        type: 200
        title: success
        detail: Adapter deleted
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
      required:
      - detail
      - title
      - type
      title: deleteadapter
      type: object
...
    standardInvalidTokenErrorResponse:
      example:
        type: 401
        title: Invalid credentials
        detail: Invalid credentials
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
      title: standardInvalidTokenErrorResponse
    standardResourceNotFoundErrorResponse:
      example:
        type: 404
        title: Not Found
        detail: Resource Not Found
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
      title: standardResourceNotFoundErrorResponse
    adapterBadRequestErrorResponse:
      example:
        type: 400
        title: Bad Request Data
        detail: Missing/Invalid parameters
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
      title: adapterBadRequestErrorResponse
...